moved token check reducer to own file to fix issue with authentication
