+++ src/redux/reducers/user.reducer.js
@@ -1,5 +1,3 @@
-import { combineReducers } from "redux";
-
 const userReducer = (state = {}, action) => {
   switch (action.type) {
     case "SET_USER":
@@ -11,18 +9,6 @@ const userReducer = (state = {}, action) => {
   }
 };
 
-const tokenCheck = (state = 'false', action) => {
-  switch (action.type) {
-    case "SET_TOKEN_CHECK":
-      return String(action.payload);
-    default:
-      return state;
-  }
-};
-
 // user will be on the redux state at:
 // state.user
-export default combineReducers({
-  userReducer,
-  tokenCheck
-})
+export default userReducer;
